pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omega-hard-print"
version = "0.1.0"
authors = [
  { name="Erik Bergsten", email="erikbergsten94@gmail.com" },
]
description = "Make PDFs from markdown"
requires-python = ">=3.8"
dependencies = [
  "weasyprint",
  "markdown-it-py",
]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
vt = "omega_hard_print.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["omega_hard_print"]

